refactor: add path aliases for all components, delete sass, --project => -p for tsc
