+++ src/app/routes/AuthorizedRoutes.ts
@@ -1,8 +1,8 @@
 import type { FC } from 'react';
-import { AuthorizedPage } from '../../pages/Authorized/AuthorizedPage.tsx';
-import ProjectCatalogPage from '../../pages/ProjectCatalog/ProjectCatalogPage.tsx';
-import ProfilePage from '../../pages/Profile/ProfilePage.tsx';
-import { ROUTES } from '../../shared/config/routes.ts';
+import { AuthorizedPage } from '@/pages/Authorized/AuthorizedPage.tsx';
+import ProjectCatalogPage from '@/pages/ProjectCatalog/ProjectCatalogPage.tsx';
+import ProfilePage from '@/pages/Profile/ProfilePage.tsx';
+import { ROUTES } from '@/shared/config/routes.ts';
 
 export const AuthorizedRoutes: {
 	path: string;
