+++ src/app/routes/PublicRoutes.ts
@@ -1,6 +1,6 @@
 import type { FC } from 'react';
-import { ROUTES } from '../../shared/config/routes.ts';
-import { WelcomePage } from '../../pages/Welcome/WelcomePage.tsx';
+import { ROUTES } from '@/shared/config/routes.ts';
+import { WelcomePage } from '@/pages/Welcome/WelcomePage.tsx';
 
 export const PublicRoutes: {
 	path: string;
