+++ src/app/routes/UnAuthorizedRoutes.ts
@@ -1,7 +1,7 @@
 import type { FC } from 'react';
-import { LoginPage } from '../../pages/Login/LoginPage.tsx';
-import { RegisterPage } from '../../pages/Register/RegisterPage.tsx';
-import { ROUTES } from '../../shared/config/routes.ts';
+import { LoginPage } from '@/pages/Login/LoginPage.tsx';
+import { RegisterPage } from '@/pages/Register/RegisterPage.tsx';
+import { ROUTES } from '@/shared/config/routes.ts';
 
 export const UnAuthorizedRoutes: {
 	path: string;
