+++ src/app/store/index.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
-import userReducer from '../../entities/User/model/userSlice';
-import themeReducer from '../../entities/Theme/model/themeSlice';
-import mobileMenuReducer from '../../entities/MobileMenu/mobileMenuSlice';
+import userReducer from '@/entities/User/model/userSlice';
+import themeReducer from '@/entities/Theme/model/themeSlice';
+import mobileMenuReducer from '@/entities/MobileMenu/mobileMenuSlice';
 
 export const store = configureStore({
 	reducer: {
