--- src/entities/User/api/index.ts
@@ -1,1 +1,0 @@
-export * from './userAPI';
