+++ src/entities/User/index.ts
@@ -1,2 +1,12 @@
-export * from './model';
-export * from './api';
+export {
+	type LoginData,
+	type RegisterData,
+	type User,
+	login,
+	logout,
+	setLoading,
+	setError,
+	registerSuccess,
+} from './model/userSlice';
+export { ValidationUtils, MIN_PASSWORD_LENGTH } from './model/constants';
+export { registerUser, loginUser, checkEmailAvailability } from './api/userAPI';
