+++ src/features/LogoutButton/LogoutButton.tsx
@@ -1,8 +1,8 @@
 import { Button } from 'antd';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../../shared/lib/hooks/useAuth.ts';
+import { useAuth } from '@/shared/lib/hooks/useAuth.ts';
 import { LogoutOutlined } from '@ant-design/icons';
-import { ROUTES } from '../../shared/config/routes.ts';
+import { ROUTES } from '@/shared/config/routes.ts';
 
 const LogoutButton = () => {
 	const navigate = useNavigate();
