+++ vite.config.ts
@@ -1,5 +1,6 @@
 import tailwindcss from '@tailwindcss/vite';
 import react from '@vitejs/plugin-react';
+import { resolve } from 'path';
 import { defineConfig } from 'vite';
 
 // https://vite.dev/config/
@@ -15,6 +16,11 @@ export default defineConfig(({ mode }) => {
 			}),
 			tailwindcss(),
 		],
+		resolve: {
+			alias: {
+				'@': resolve(__dirname, './src'),
+			},
+		},
 		base: isProduction ? '/' : '/',
 		build: {
 			outDir: 'dist',
